+++ src/components/abstract-component.js
@@ -1,0 +1,18 @@
+import {createElement} from "./utils";
+
+export class AbstractComponent {
+  constructor() {
+    this._element = null;
+  }
+
+  getElement() {
+    if (!this._element) {
+      this._element = createElement(this.getTemplate());
+    }
+
+    return this._element;
+  }
+  getTemplate() {
+    throw Error(`Abstract method not implemented`);
+  }
+}
