+++ src/components/board.js
@@ -1,5 +1,11 @@
-export const createBoardTemplate = () => {
-  return `<section class="board container">
-      <div class="board__tasks"></div>
+import {AbstractComponent} from './abstract-component';
+
+export class Board extends AbstractComponent {
+  constructor() {
+    super();
+  }
+  getTemplate() {
+    return `<section class="board container">
     </section>`;
-};
+  }
+}
