--- src/components/filter-sec.js
@@ -1,4 +1,0 @@
-export const createFilterSection = () =>
-  `<section class="main__filter filter container">
-  </section>`;
-
