+++ src/components/filter.js
@@ -1,14 +1,22 @@
-export const createFilterTemplate = (filter) =>
-  `<input
+import {AbstractComponent} from './abstract-component';
+
+export class Filter extends AbstractComponent {
+  constructor(filters) {
+    super();
+    this._filters = filters;
+  }
+  getTemplate() {
+    return `<section class="main__filter filter container">
+    ${this._filters.map((filter) => `<input
         type="radio"
-        id="filter__all"
+        id="filter__${filter.title}"
         class="filter__input visually-hidden"
-        name="filter"
-        checked
-        ${filter.count === 0 ? `disabled` : ``}
+        name="filter" ${filter.count === 0 ? `disabled` : ``}
       />
-      <label for="filter__all" class="filter__label">
-    ${filter.title}
-      <span class="filter__all-count">${filter.count}</span></label>
-   `;
+    <label for="filter__${filter.title}" class="filter__label">
+      ${filter.title}<span class="filter__all-count"> ${filter.count}</span></label
+    >`).join(``)}
+      </section>`.trim();
+  }
+}
