+++ src/components/load-more-button.js
@@ -1,3 +1,10 @@
-export const createLoadMoreButtonTemplate = () => {
-  return `<button class="load-more" type="button">load more</button>`;
-};
+import {AbstractComponent} from './abstract-component';
+
+export class LoadMoreButton extends AbstractComponent {
+  constructor() {
+    super();
+  }
+  getTemplate() {
+    return `<button class="load-more" type="button">load more</button>`;
+  }
+}
