+++ src/components/sorting.js
@@ -1,7 +1,14 @@
-export const createSortingTemplate = () => {
-  return `<div class="board__filter-list">
+import {AbstractComponent} from './abstract-component';
+
+export class Sort extends AbstractComponent {
+  constructor() {
+    super();
+  }
+  getTemplate() {
+    return `<div class="board__filter-list">
       <a href="#" class="board__filter">SORT BY DEFAULT</a>
       <a href="#" class="board__filter">SORT BY DATE up</a>
       <a href="#" class="board__filter">SORT BY DATE down</a>
     </div>`;
-};
+  }
+}
