+++ src/components/task-list.js
@@ -1,0 +1,7 @@
+import {AbstractComponent} from './abstract-component';
+
+export class TaskList extends AbstractComponent {
+  getTemplate() {
+    return `<div class="board__tasks"></div>`;
+  }
+}
